@@ -4,14 +4,13 @@ apply plugin: 'com.github.dcendents.android-maven'
 group='com.github.noelchew'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 29
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 27
-        versionCode 36
-        versionName "0.1.36"
+        targetSdkVersion 29
+        versionCode 39
+        versionName "0.1.39"
     }
     buildTypes {
         release {
@@ -25,15 +24,14 @@ dependencies {
     
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     testImplementation 'junit:junit:4.12'
-    implementation "com.android.support:support-v4:${rootProject.ext.googleVersion}"
-    implementation "com.android.support:appcompat-v7:${rootProject.ext.googleVersion}"
-    compileOnly 'com.github.renaudcerrato:static-maps-api:1.0.4'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
 
     // https://mvnrepository.com/artifact/com.google.code.gson/gson
 //    implementation group: 'com.google.code.gson', name: 'gson', version: '2.2.4'
 
     // https://mvnrepository.com/artifact/com.google.code.gson/gson
-    implementation group: 'com.google.code.gson', name: 'gson', version: '2.8.0'
+    implementation group: 'com.google.code.gson', name: 'gson', version: '2.8.6'
 
 
 
