@@ -1,15 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "com.noelchew.ncutils.demo"
         minSdkVersion 16
-        targetSdkVersion 27
-        versionCode 36
-        versionName "0.1.36"
+        targetSdkVersion 29
+        versionCode 39
+        versionName "0.1.39"
     }
     buildTypes {
         release {
@@ -23,10 +22,10 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(':library')
     testImplementation 'junit:junit:4.12'
-    implementation "com.android.support:support-v4:${rootProject.ext.googleVersion}"
-    implementation "com.android.support:appcompat-v7:${rootProject.ext.googleVersion}"
-    implementation "com.android.support:recyclerview-v7:${rootProject.ext.googleVersion}"
-    implementation "com.android.support:design:${rootProject.ext.googleVersion}"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'com.google.android.material:material:1.1.0'
     implementation 'com.joanzapata.iconify:android-iconify-material:2.2.2' // (v2.0.0)
     implementation 'com.joanzapata.iconify:android-iconify-material-community:2.2.2' // (v1.4.57)
     implementation 'com.jude:easyrecyclerview:4.4.2'
